Extract logic in documentNavigator into a core module
